pyproject.toml
[tool.poetry]
name = "task4"
version = "0.1.0"
description = ""
authors = ["python_candidate1 <selver.said01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
Pygments = "^2.10.0"
requests = "^2.26.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
